requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blis==0.7.9
catalogue==2.0.8
cchardet==2.1.7
celery==5.2.7
celery-progress==0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
confection==0.0.4
configparser==5.3.0
crispy-bootstrap5==0.7
cymem==2.0.7
Cython==0.29.34
Django==4.1.7
django-celery-results==2.5.0
django-crispy-forms==2.0
elastic-transport==8.4.0
elasticsearch==8.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fake-useragent==1.1.3
filelock==3.12.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
googletransx==2.4.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
prompt-toolkit==3.0.38
pycares==4.3.0
pycparser==2.21
pydantic==1.10.7
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.2.0
pytz==2022.7.1
redis==4.5.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
schedule==1.2.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@786815dd05681e2421cd03aa9acf5ab5c773bce9
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.10
threadpoolctl==3.1.0
tqdm==4.64.1
tweepy==4.12.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.14
vine==5.0.0
wasabi==1.1.1
wcwidth==0.2.6
yarl==1.9.1
